pyproject.toml
[tool.poetry]
name = "main-service"
version = "0.1.0"
description = ""
authors = ["Mihail Pugachev <xd393727@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
httpx = "^0.27.0"
pytest = "^8.2.0"
grpcio = "^1.63.0"
grpcio-tools = "^1.63.0"
faker = "^25.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
